Split each office-hours Cloud Function producer into a firebase-functions-free <name>-core.ts (the DI core fn, pure helpers, GraphQL query strings, wire types) re-exported by the onSchedule wrapper, so the new office-hours-snapshot producer can import the cores as libraries without triggering firebase-functions param registration. Behavior-preserving; existing functions tests pass unchanged via re-exports.
